backend/ImhotepChef/requirements.txt
Django==5.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg[binary]==3.2.3
django-cors-headers==4.3.1
setuptools